requirements.txt
requests-oauthlib~=1.3.0
transformers~=4.3.2
fugashi~=1.1.0
ipadic~=1.0.0
tensorflow~=2.4.1
tf-models-official~=2.4.0
pandas~=1.0.3
numpy~=1.18.2